feat(cmd/gpud): require sudo for join, up, down, login
